sqlite_version

+++ src/main/java/com/alura/hotelalura/springboottres/controller/responses/ConsultarResponses.java
@@ -1,12 +1,12 @@
 package com.alura.hotelalura.springboottres.controller.responses;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 public record ConsultarResponses(
      LocalDate checkIn,
      LocalDate checkOut,
      String categoria,
-     BigDecimal precio
+     Float precio
 )
 {}
