+++ src/main/java/com/alura/hotelalura/springboottres/persitence/dto/habitacion_tipo/HabitacionTipoModel.java
@@ -1,9 +1,9 @@
 package com.alura.hotelalura.springboottres.persitence.dto.habitacion_tipo;
-import java.math.BigDecimal;
+
 
 
 public record HabitacionTipoModel(
     String categoria,
-    BigDecimal percioUnitario,
+    Float percioUnitario,
     Integer puntosUnitario
 ){}
