+++ src/main/java/com/alura/hotelalura/springboottres/persitence/dto/reserva/ListarReservacion.java
@@ -1,13 +1,12 @@
 package com.alura.hotelalura.springboottres.persitence.dto.reserva;
 
-import java.math.BigDecimal;
 
 import com.alura.hotelalura.springboottres.persitence.entity.ReservaEntity;
 
 public record ListarReservacion(
     String reservaId,
     Integer habitacion,
-    BigDecimal precio
+    Float precio
 ) 
 {
     public ListarReservacion(ReservaEntity entity)
